third_party/blink/web_tests/platform/generic/http/tests/security/contentSecurityPolicy/cached-frame-csp-expected.txt

CONSOLE ERROR: Refused to execute inline script because it violates the following Content Security Policy directive: "script-src 'none'". Either the 'unsafe-inline' keyword, a hash ('sha256-HNo4gm6q1snjil9OgSmpbSnn3Bhbw4R+r+SBLYD3ENc='), or a nonce ('nonce-...') is required to enable inline execution.

CONSOLE ERROR: Refused to execute inline script because it violates the following Content Security Policy directive: "script-src 'none'". Either the 'unsafe-inline' keyword, a hash ('sha256-HNo4gm6q1snjil9OgSmpbSnn3Bhbw4R+r+SBLYD3ENc='), or a nonce ('nonce-...') is required to enable inline execution.

Check that a CSP header added by a 304 response does not override one from the original request.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Check that the nonce is the same, meaning that the document was only generated once:
PASS frame1.contentDocument.querySelector("input").value == frame2.contentDocument.querySelector("input").value is true
PASS successfullyParsed is true

TEST COMPLETE

 
